+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] v0.2.0
+## [Unreleased] v0.1.3
 
 ## 0 Month 2021
 ### Added
@@ -15,12 +15,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+## Released [v0.1.2]
+## 7 Jan 2021
+
+### Changed
+- added zap logger to strings hash method
+- httplib interceptors to use zap logger
+
 
 ## Released [v0.1.1]
 ## 7 Jan 2021
+
 ### Added
 - moved pkg from private repos to public utils
 
 
-[Unreleased]: https://github.com/rovergulf/utils/compare/v0.1.1...HEAD
+[Unreleased]: https://github.com/rovergulf/utils/compare/v0.1.2...HEAD
+[Unreleased]: https://github.com/rovergulf/utils/compare/v0.1.2...v0.1.1
 [v0.1.1]: https://github.com/rovergulf/utils/tree/v0.1.1